🔧 (distiller.py, informant.py, language_processor.py, listen.py): Change logger level from info to debug for non-critical log messages The logger level was changed from info to debug for non-critical log messages to reduce the verbosity of the logs and make it easier to identify important information. This change will help in maintaining a cleaner and more readable log, which is crucial for effective debugging and monitoring.
✨ (boot.launch.py): Add new vision capable lpu node with model llava and base_url
✨ (boot.launch.py): Add new distill node named the_spokesman with action_server_name instruct
♻️ (boot.launch.py): Change update_interval of sensation node to 0.0 for instantaneous updates
♻️ (boot.launch.py): Change update_interval of song node to 15.0 for more frequent updates
🐛 (memorialist.py, sense.py, stream.py, transcribe.py): Change logger level from info to debug for less verbose logging
The new vision capable lpu node and the_spokesman node are added to enhance the functionality of the application. The update_interval of sensation and song nodes are adjusted to improve the responsiveness of the application. The logger level is changed from info to debug in several python files to reduce the verbosity of the logs, making it easier to identify and troubleshoot issues.

♻️ (play_song.py): Refactor song_callback function to improve code readability and efficiency
💡 (play_song.py): Change log level from info to debug for song reception
The song_callback function was refactored to improve code readability and efficiency. The publishing of the defined song was removed as it was redundant and the song to be played is now directly published, reducing unnecessary steps. The log level for song reception was changed from info to debug to reduce log clutter during normal operation, while still providing detailed information for debugging purposes.